requirements.txt
Flask==1.1.2
Flask-SQLAlchemy==2.5.1
PyMySQL==1.0.2
python-dateutil==2.8.1
python-dotenv==0.16.0